* added 3 web locations for program library: WEB_APP_DIR (emergent) WEB_SCI_DIR (CCN),   WEB_USER_DIR (user-settable) * updated ProgLib to emergent std coding conventions, using ProgLib_List, etc, * added PRE_CALL_InitProgLib to initialize program library as needed.. * when saving a program to wiki prog lib, sets doc links to that wiki page * removed current_key_bindings as a user-settable setting * first-pass conversion of so * fixed PVLV VTA VSpatch code to not use layers if gains = 0

+++ src/temt/ta_gui/ISelectable.cpp
@@ -258,7 +258,7 @@ void ISelectable::FillContextMenu_EditItems_impl(taiWidgetActions* menu,
   if (ea == 0)
     return;
   
-  int kb_set = (static_cast<int>(taMisc::current_key_bindings));
+  int kb_set = taMisc::current_key_bindings; // enums are automatically int's right?  don't need a static cast here..
 
   if (menu->count() > 0)
     menu->AddSep();
